+++ lib/src/transport/local_transport.dart
@@ -1,7 +1,10 @@
 import 'dart:async';
+import 'dart:collection';
 
-import '../message/abstract_message.dart';
-import 'abstract_transport.dart';
+import 'package:connectanum/authentication.dart';
+import 'package:connectanum/connectanum.dart';
+import 'package:connectanum/src/message/hello.dart';
+import 'package:connectanum/src/message/welcome.dart';
 
 class LocalTransport extends AbstractTransport {
   final _receiveController = StreamController<AbstractMessage?>.broadcast();
@@ -11,7 +14,13 @@ class LocalTransport extends AbstractTransport {
   final Completer _onDisconnect = Completer<void>();
   final Completer _onConnectionLost = Completer<void>();
 
+  String authenticationPassword;
+  Hello? _hello;
+  AbstractAuthentication? _authentication;
+  String? _signature;
   bool _isOpen = false;
+
+  LocalTransport({this.authenticationPassword = "password"});
 
   /// Allow tests to listen to what was sent via the transport
   Stream<AbstractMessage> get sentMessages => _sentMessagesController.stream;
@@ -52,12 +61,76 @@ class LocalTransport extends AbstractTransport {
   }
 
   @override
-  Stream<AbstractMessage?>? receive() {
+  Stream<AbstractMessage?> receive() {
     return _receiveController.stream;
   }
 
   @override
   void send(AbstractMessage message) {
+    if (message is Hello) {
+      var craAuthentication = CraAuthentication(authenticationPassword);
+      var scramAuthentication = ScramAuthentication(authenticationPassword);
+      if (message.details.authmethods?.contains(craAuthentication.getName()) ?? false) {
+        var extra = Extra(
+            salt: message.details.salt ?? 'salt',
+            keyLen: 32,
+            iterations: 1000,
+            challenge: '{"authid":"${message.details.authid}","authrole":"client","authmethod":"${craAuthentication.getName()}","authprovider":"local","nonce":"local","timestamp":"1970-01-01T12:00Z","session":1}'
+        );
+        craAuthentication.challenge(extra).then((authenticate) {
+          _hello = message;
+          _signature = authenticate.signature;
+          _authentication = craAuthentication;
+          _receiveController.add(Challenge(craAuthentication.getName(), extra));
+        });
+      } else if (message.details.authmethods?.contains(scramAuthentication.getName()) ?? false) {
+        var extra = Extra(
+            iterations: 1,
+            memory: 100,
+            salt: 'AQ==',
+            nonce: '${message.details.authextra?['nonce']}AQ==',
+            kdf: ScramAuthentication.kdfArgon);
+        var authExtra = HashMap<String, Object?>();
+        authExtra['nonce'] = extra.nonce;
+        authExtra['channel_binding'] = null;
+        authExtra['cbind_data'] = null;
+        _hello = message;
+        _authentication = scramAuthentication;
+        _signature = scramAuthentication.createSignature(
+            message.details.authid ?? '',
+            message.details.authextra?['nonce'],
+            extra,
+            authExtra);
+        _receiveController.add(Challenge(scramAuthentication.getName(), extra));
+      } else {
+        _receiveController.add(Welcome(1, Details.forWelcome(
+          authId: _hello?.details.authid,
+          authMethod: _authentication?.getName(),
+          authProvider: 'local',
+          authRole: 'client',
+          realm: _hello?.realm,
+        )));
+      }
+    } else if (message is Authenticate) {
+      bool success = false;
+      if (_authentication is CraAuthentication) {
+        success = message.signature == _signature;
+      }
+      if (_authentication is ScramAuthentication) {
+        success = message.signature == _signature;
+      }
+      if (success) {
+        _receiveController.add(Welcome(1, Details.forWelcome(
+          authId: _hello?.details.authid,
+          authMethod: _authentication?.getName(),
+          authProvider: 'local',
+          authRole: 'client',
+          realm: _hello?.realm,
+        )));
+      } else {
+        _receiveController.add(Abort("Wrong password", message: "Authentication process failed!"));
+      }
+    }
     _sentMessagesController.add(message);
   }
 }